Draw Time Series Plot for Blood Pressure

Draw Time Series Plot for Blood Pressure

Usage

drawBloodPressure(
  adlb,
  settings,
  labs = c("Diastolic Blood Pressure", "Systolic Blood Pressure")
)

Arguments

adlb

lab data in tall format that must contain DY for study day, VISITN for visit number, TEST for lab test, and STRESN for lab value

settings

settings object with column mappings

labs

character string or character vector specifying which labs from TEST to include

Value

ggplot object