Draw Time Series Plot for Fold Change from ULN

Draw Time Series Plot for Fold Change from ULN

Usage

drawULNFoldChange(
  adlb,
  settings,
  labs = c("Bicarbonate", "Blood Urea Nitrogen", "Calcium", "Chloride", "Phosphorus",
    "Potassium", "Sodium")
)

Arguments

adlb

lab data in tall format that must contain DY for study day, VISITN for visit number, TEST for lab test, and STRESN for lab value

settings

settings object with column mappings

labs

character string or character vector specifying which labs from TEST to include

Value

ggplot object